pyproject.toml
[project]
name = "remind-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.7",
    "beautifulsoup4>=4.13.4",
    "discord-py>=2.5.2",
    "lxml>=5.4.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.3",
]
